% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capstoneproject.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Use the Use_vignette once to generate the vignette folder in the package's skeleton where the vignette documentation
made with rmarkdown will be made
use_vignette("capstoneproject")
Use the use_testthat() once to generate the test folder in the package's skeleton where the test scripts will be placed}
\usage{
eq_clean_data()
}
\value{
a dataframe which contains the Eathquake data filtered required for mapping in a timeline the data
}
\description{
use_testthat()
}
\details{
Funcion for reading and cleaning the Earthquake data
}

