% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capstoneproject.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Funcion for title case the Earthquake's Location Data-Name}
\usage{
eq_location_clean(datfram)
}
\arguments{
\item{datfram}{is the dataframe that contains location names written in Uper case}
}
\value{
a dataframe which contains the Eathquake data filtered required for mapping in a timeline the data and the Tittle Case Location
}
\description{
Funcion for title case the Earthquake's Location Data-Name
}

