% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capstoneproject.R
\name{eq_map}
\alias{eq_map}
\title{Funcion for plotting in an interactive map the location of a set of Earthquakes ocurred in a given country}
\usage{
eq_map(country = "MEXICO", annot_col = "date", dat = "1980")
}
\arguments{
\item{annot_col}{is the annotation to be displayed using a pop up use date to display only the date or blank for enhanced data display}

\item{dat}{is the starting year from where the data will be displayed}

\item{Country}{is the Country that contains the Earthquake's data to be displayed}
}
\description{
Funcion for plotting in an interactive map the location of a set of Earthquakes ocurred in a given country
}
\examples{
\dontrun{
eq_map()
}

}

