% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capstoneproject.R
\name{eq_map_simple}
\alias{eq_map_simple}
\title{Funcion for plotting in a simple map where a set of Earthquakes occured}
\usage{
eq_map_simple(country = "MEXICO", annot_col = "DATE", date = "1980")
}
\arguments{
\item{annot_col}{is the annotation to be displayed using a pop up}

\item{Country}{is the Country that contains the Earthquake's data to be displayed}

\item{year}{is the starting year from where the data will be displayed}
}
\description{
Funcion for plotting in a simple map where a set of Earthquakes occured
}

