% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capstoneproject.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Funcion for ploting an Earthquake's Location timeline}
\usage{
geom_timeline(datfram, start_date = "1980-01-01", end_date = "2014-01-01",
  countries = c("MEXICO", "USA", "JORDAN"))
}
\arguments{
\item{datfram}{is the dataframe that contains the Earthquake's data}

\item{start_date}{is the starting date from when the Earthquakes are going to be ploted}

\item{end_date}{is the end date up to the Earthquakes are going to be ploted}

\item{countries}{is a vector which contains the list of Countries that are going to be plotted}
}
\description{
Funcion for ploting an Earthquake's Location timeline
}

