% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capstoneproject.R
\docType{data}
\name{GeomTimeLineAnnotation}
\alias{GeomTimeLineAnnotation}
\title{Funcion for ploting an Earthquake's Location timeline building a GEOM Function from scratch but it will include annotations}
\format{An object of class \code{GeomTimeLineAnnotation} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeLineAnnotation
}
\description{
The GeomTimeLineAnnotation will use a Dataframe compiled using the function eq_clean_data
The GeomTimeLineAnnotation function is a prototype function which will be used as foundation for our geom_timeline function
The GeomTimeLineAnnotation function will take advantage of the ggplot2's geom_point
Using the Earthquakes' dates as X-axis main values, the Y-axis value will be not relevant while plotting a timeline (horizontal bar)
The GeomTimeLineAnnotation will use a number of annotations that will be plotted (Number of Major Earthquakes to be highlighted)
The geom_point's size and colour will be defined by the Earthquake's magnitude
The GeomTimeLineAnnotation was build using the Function Prototype provided in the Course's Material 4.7.1 Building a New Geom
}
\keyword{datasets}

