% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(noaa_data)
}
\arguments{
\item{noaa_data}{a dataframe contains information about 5,933 earthquakes over an approximately 4,000 year time span}
}
\value{
clean_data new dataframe repersents a cleaned data from NOAA
}
\description{
this function used to clean the data comes from NOAA and add \code{LOCATION_NAME} column
}
\examples{
\dontrun{
noaa <- readr::read_delim("filepath")
clean_noaa_data <- eq_location_clean(noaa)
}
}
