% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{eq_map}
\alias{eq_map}
\title{eq_map}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{the data obtained from NOAA .}

\item{annot_col}{column used for annotation in the pop-up graph}
}
\description{
Function that takes a data from NOAA and annotation column to
get back a leaflet map represents individual earthquakes .
}
\examples{
\dontrun{
noaa \%>\%
 eq_clean_data() \%>\%
 dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
 eq_map(annot_col = "DATE")
}
}
