% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{look at Arguments section in  \code{ggplot2::geom_point()}}

\item{data}{look at Arguments section in  \code{ggplot2::geom_point()}}

\item{stat}{look at Arguments section in  \code{ggplot2::geom_point()}}

\item{position}{look at Arguments section in  \code{ggplot2::geom_point()}}

\item{na.rm}{look at Arguments section in  \code{ggplot2::geom_point()}}

\item{show.legend}{look at Arguments section in  \code{ggplot2::geom_point()}}

\item{inherit.aes}{look at Arguments section in  \code{ggplot2::geom_point()}}

\item{...}{look at Arguments section in  \code{ggplot2::geom_point()}}
}
\description{
shows a timeline of NOAA Significant earthquakes for an individual country in the y axis and dates in x axis
and the size of the points is associated with the earthquakes magnitude and
the color is represented the total number of deaths
}
\section{Aesthetics }{

 \itemize{
  \item{\code{x}} : the DATE column in the NOAA dataframe .
  \item{\code{size}} :  a numeric column which represents Richter scale value .
  \item{\code{color}} : a numeric column which represents the number of deaths per country .
  \item{\code{fill}} : it may be the number of deaths per country too .
  \item{\code{y}} : it is column that represents \code{COUNTRY} name .
  \item{\code{alpha}}
  \item{\code{shape}}
}
}

\examples{
\dontrun{
noaa \%>\%
eq_clean_data() \%>\%
filter(YEAR >= 2000, COUNTRY == "USA") \%>\%
ggplot(aes(x = DATE, size = as.numeric(EQ_PRIMARY), color = as.numeric(TOTAL_DEATHS))) +
geom_timeline()
}
}
