% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{look at Arguments section in  \code{ggplot2::geom_point()}}

\item{data}{look at Arguments section in  \code{ggplot2::geom_point()}}

\item{stat}{look at Arguments section in  \code{ggplot2::geom_point()}}

\item{position}{look at Arguments section in  \code{ggplot2::geom_point()}}

\item{na.rm}{look at Arguments section in  \code{ggplot2::geom_point()}}

\item{show.legend}{look at Arguments section in  \code{ggplot2::geom_point()}}

\item{inherit.aes}{look at Arguments section in  \code{ggplot2::geom_point()}}

\item{...}{look at Arguments section in  \code{ggplot2::geom_point()}}
}
\description{
Adds labels for the locations of individual earthquaks
}
\section{Aesthetics }{

\itemize{
  \item{\code{x}} : the \code{DATE} of the earthquake .
  \item{\code{label}} :  a character column which represents the location of individual earthquaks .  .
  \item{\code{color}} : a numeric column which represents the number of deaths per country .
  \item{\code{fill}} : it may be the number of deaths per country too .
  \item{\code{size}} : the magnitude of the earthquake measured on the Richter scale
}
}

\examples{
\dontrun{
eq_data <- noaa \%>\%
 eq_clean_data() \%>\%
 eq_location_clean() \%>\%
 filter(YEAR >= 2000, COUNTRY == "USA") \%>\%
 ggplot(aes(x = DATE, size = as.numeric(EQ_PRIMARY),
 color = as.numeric(TOTAL_DEATHS), label = LOCATION)) +
 geom_timeline() +
 geom_timeline_label()
}
}
